requirements.txt
yfinance>=0.2.63
alpaca-py>=0.30.1
pandas>=2.2.2
polars>=2.0.2
numpy>=1.26.4
requests>=2.32.3
httpx>=0.27.2
beautifulsoup4>=4.13.4
gdown>=5.2.0
tqdm>=4.66.5
tenacity>=8.5.0
python-dotenv>=1.0.1
pyyaml>=6.0.2
sentence-transformers>=2.2.2
transformers>=4.20.1
torch>=2.4.1
huggingface-hub>=0.25.2
scipy>=1.14.1
scikit-learn>=1.5.2
nltk>=3.9.1
matplotlib>=3.5.2
seaborn>=0.13.2
pydantic>=2.9.2